refactor: extract invoice type (#98016)
